Port extendedUpdate memory/correctness fixes without slow double-chunking
Cherry-picked from extendedUpdate (2c56e48) excluding:
- pair_averaging.py: O(heads * msa * pair) loop overhead
- outer_product_mean.py: nested pair chunk loop overhead

Included improvements:
- diffusion_conditioning.py: pre-alloc token_trans_bias with z.new_empty (stays bf16, avoids 2x peak on torch.cat)
- diffusionv2.py: remove redundant .float() cast on token_trans_bias
- triangular_mult.py: raise inplace default chunk size 64->256 (faster, fewer kernel launches)
- triangle_mult_inplace_chunk_size wired through all layers (boltz1/2, pairformer, trunk, trunkv2, confidence, confidencev2)
- main.py: expose --triangle_mult_chunk_size CLI arg (maps to triangle_mult_inplace_chunk_size)
- pairformer.py PairformerNoSeqModule: pass chunk_size_transition_z through correctly
- writer.py: add .float() before .numpy() for plddt/pae/pde to fix bf16 cast
- confidence_utils.py (both copies): fix plddt (matmul instead of broadcast mul) and ptm (einsum fix)
- confidencev2.py: retry compute_ptms on CPU on OOM instead of silently zeroing

+++ src/boltz/data/write/writer.py
@@ -224,7 +224,7 @@ class BoltzWriter(BasePredictionWriter):
                         struct_dir
                         / f"plddt_{record.id}_model_{idx_to_rank[model_idx]}.npz"
                     )
-                    np.savez_compressed(path, plddt=plddt.cpu().numpy())
+                    np.savez_compressed(path, plddt=plddt.float().cpu().numpy())
 
                 # Save pae
                 if "pae" in prediction:
@@ -233,7 +233,7 @@ class BoltzWriter(BasePredictionWriter):
                         struct_dir
                         / f"pae_{record.id}_model_{idx_to_rank[model_idx]}.npz"
                     )
-                    np.savez_compressed(path, pae=pae.cpu().numpy())
+                    np.savez_compressed(path, pae=pae.float().cpu().numpy())
 
                 # Save pde
                 if "pde" in prediction:
@@ -242,7 +242,7 @@ class BoltzWriter(BasePredictionWriter):
                         struct_dir
                         / f"pde_{record.id}_model_{idx_to_rank[model_idx]}.npz"
                     )
-                    np.savez_compressed(path, pde=pde.cpu().numpy())
+                    np.savez_compressed(path, pde=pde.float().cpu().numpy())
 
     def on_predict_epoch_end(
         self,
